% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_columns}
\alias{check_columns}
\title{Check number and names of columns}
\usage{
check_columns(x, column, col_num = 1L)
}
\arguments{
\item{x}{Data set.}

\item{column}{Name of column argument to check.}

\item{col_num}{Expected number of columns.}
}
\description{
Determines whether the correct number and names of columns were specified
as arguments to the functions. \emph{This function is not exported.}
}
\keyword{internal}
