% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duration.R
\name{check_duration}
\alias{check_duration}
\title{Title}
\usage{
check_duration(
  .data,
  duration_col = "Duration (in seconds)",
  min = NULL,
  max = NULL,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
