% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exclusions.R
\name{check_exclusions}
\alias{check_exclusions}
\title{Title}
\usage{
check_exclusions(
  .data,
  exclusion_types = c("duplicates", "duration", "ip", "language", "location",
    "preview", "progress", "resolution"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
