% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_language.R
\name{check_language}
\alias{check_language}
\title{Title}
\usage{
check_language(
  .data,
  language = "EN",
  lang_col = "UserLanguage",
  print_tibble = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
