% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_location.R
\name{check_location}
\alias{check_location}
\title{Check data rows where location information is absent and outside of the the US}
\usage{
check_location(
  .data,
  location_col = c("LocationLatitude", "LocationLongitude"),
  include_na = FALSE,
  print_tibble = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{.data}{Data frame (preferably directly from Qualtrics imported using {qualtRics}.)}

\item{location_col}{Two element vector specifying columns for latitude and longitude (in that order).}

\item{include_na}{Logical indicating whether to include rows with NA in latitude and longitude columns in the output list of potentially excluded data.}

\item{print_tibble}{Logical indicating whether to print the tibble to the console.}

\item{quiet}{Logical indicating whether to print message to console.}
}
\value{
The output is a data frame of the rows that are located outside of the US and (if \code{include_na == FALSE}) rows with no location information. For a function that excludes these rows, use \link{exclude_location}.
}
\description{
Check data rows where location information is absent and outside of the the US
}
