% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_preview.R
\name{check_preview}
\alias{check_preview}
\title{Title}
\usage{
check_preview(
  .data,
  preview_col = "Status",
  print_tibble = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
