% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_resolution.R
\name{check_resolution}
\alias{check_resolution}
\title{Title}
\usage{
check_resolution(
  .data,
  width_min = 1000,
  height_min = 0,
  res_col = "Resolution",
  print_tibble = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
