% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_exclusions.R
\name{collapse_exclusions}
\alias{collapse_exclusions}
\title{Title}
\usage{
collapse_exclusions(
  .data,
  exclusion_types = c("duplicates", "duration", "ip", "location", "progress")
)
}
\arguments{
\item{exclusion_types}{}
}
\value{

}
\description{
Title
}
