% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude.R
\name{exclude}
\alias{exclude}
\title{Title}
\usage{
exclude(
  .data,
  exclusion_types = c("duplicates", "duration", "ip", "language", "location",
    "preview", "progress", "resolution"),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
