% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_duplicates.R
\name{exclude_duplicates}
\alias{exclude_duplicates}
\title{Title}
\usage{
exclude_duplicates(
  .data,
  ip_col = "IPAddress",
  location_col = c("LocationLatitude", "LocationLongitude"),
  check_ip = TRUE,
  check_location = TRUE,
  include_na = FALSE,
  id_col = "ResponseId",
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
