% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_ip.R
\name{exclude_ip}
\alias{exclude_ip}
\title{Title}
\usage{
exclude_ip(
  .data,
  ip_col = "IPAddress",
  country = "US",
  id_col = "ResponseId",
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
