% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_location.R
\name{exclude_location}
\alias{exclude_location}
\title{Exclude data rows where location information is absent and outside of the the US}
\usage{
exclude_location(
  .data,
  location_col = c("LocationLatitude", "LocationLongitude"),
  include_na = FALSE,
  id_col = "ResponseId",
  quiet = FALSE
)
}
\arguments{
\item{.data}{Data frame (preferably directly from Qualtrics imported using {qualtRics}.)}

\item{location_col}{Two element vector specifying columns for latitude and longitude (in that order).}

\item{include_na}{Logical indicating whether to include rows with NA in latitude and longitude columns in the output list of potentially excluded data.}

\item{id_col}{Column label of column specifying unique row ID. This used to exclude from \code{.data} the rows outputted from \code{\link[=check_location]{check_location()}}.}

\item{quiet}{Logical indicating whether to print results to console.}
}
\value{
The output is a data frame of the rows that are located inside the US and (if \code{include_na == FALSE}) rows with location information. For a function that creates a data frame that outputs the excluded rows, use \code{\link[=check_location]{check_location()}}.
}
\description{
Exclude data rows where location information is absent and outside of the the US
}
