% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_progress.R
\name{exclude_progress}
\alias{exclude_progress}
\title{Title}
\usage{
exclude_progress(
  .data,
  min = 100,
  finished_col = "Finished",
  progress_col = "Progress",
  id_col = "ResponseId",
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
