% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{exclude_resolution}
\alias{exclude_resolution}
\title{Exclude unacceptable screen resolution}
\usage{
exclude_resolution(x, print = FALSE, silent = FALSE, ...)
}
\arguments{
\item{x}{Data frame (preferably imported from Qualtrics using \{qualtRics\}).}

\item{print}{Logical indicating whether to print returned tibble to
console.}

\item{silent}{Logical indicating whether to print message to console. Note
this argument controls the exclude message not the check message.}

\item{...}{Inherit parameters from \code{\link[=mark_resolution]{mark_resolution()}}.}
}
\value{
An object of the same type as \code{x} that excludes rows
that have unacceptable screen resolutions.
For a function that checks for these rows, use \code{\link[=check_resolution]{check_resolution()}}.
For a function that marks these rows, use \code{\link[=mark_resolution]{mark_resolution()}}.
}
\description{
The \code{exclude_resolution()} function removes
rows that have unacceptable screen resolution.
The function is written to work with data from
\href{https://www.qualtrics.com/}{Qualtrics} surveys.
}
\details{
Default column names are set based on output from the
\href{https://docs.ropensci.org/qualtRics/reference/fetch_survey.html}{\code{qualtRics::fetch_survey()}}.

The function outputs to console a message about the number of rows
with unacceptable screen resolution.
}
\examples{
# Exclude low screen resolutions
data(qualtrics_text)
df <- exclude_resolution(qualtrics_text)

# Remove preview data first
df <- qualtrics_text \%>\%
  exclude_preview() \%>\%
  exclude_resolution()
}
\seealso{
Other resolution functions: 
\code{\link{check_resolution}()},
\code{\link{mark_resolution}()}

Other exclude functions: 
\code{\link{exclude_duplicates}()},
\code{\link{exclude_duration}()},
\code{\link{exclude_ip}()},
\code{\link{exclude_location}()},
\code{\link{exclude_preview}()},
\code{\link{exclude_progress}()}
}
\concept{exclude functions}
\concept{resolution functions}
