% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_marked_column}
\alias{keep_marked_column}
\title{Keep column with marked rows}
\usage{
keep_marked_column(x, column, keep)
}
\arguments{
\item{x}{Data set.}

\item{column}{Name of exclusion column.}

\item{keep}{Logical indicating whether to keep or remove exclusion column.}
}
\description{
For check_*() functions, keep the column that has marked rows and move to
first column or remove the column depending on \code{keep} flag.
\emph{This function is not exported.}
}
\keyword{internal}
