% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_duration.R
\name{mark_duration}
\alias{mark_duration}
\title{Mark minimum or maximum durations}
\usage{
mark_duration(
  x,
  min_duration = 10,
  max_duration = NULL,
  id_col = "ResponseId",
  duration_col = "Duration (in seconds)",
  quiet = FALSE
)
}
\arguments{
\item{x}{Data frame (preferably imported from Qualtrics using \{qualtRics\}).}

\item{min_duration}{Minimum duration that is too fast in seconds.}

\item{max_duration}{Maximum duration that is too slow in seconds.}

\item{id_col}{Column name for unique row ID (e.g., participant).}

\item{duration_col}{Column name for durations.}

\item{quiet}{Logical indicating whether to print message to console.}
}
\value{
An object of the same type as \code{x} that includes a column marking rows
with fast and slow duration.
For a function that checks for these rows, use \code{\link[=check_duration]{check_duration()}}.
For a function that excludes these rows, use \code{\link[=exclude_duration]{exclude_duration()}}.
}
\description{
The \code{mark_duration()} function creates a column labeling
rows with fast and/or slow duration.
The function is written to work with data from
\href{https://www.qualtrics.com/}{Qualtrics} surveys.
}
\details{
Default column names are set based on output from the
\href{https://docs.ropensci.org/qualtRics/reference/fetch_survey.html}{\code{qualtRics::fetch_survey()}}.
By default, minimum durations of 10 seconds are checked, but either
minima or maxima can be checked with the \code{min_duration} and
\code{max_duration} arguments. The function outputs to console separate
messages about the number of rows that are too fast or too slow.

This function returns the fast and slow rows.
}
\examples{
# Mark durations faster than 100 seconds
data(qualtrics_text)
df <- mark_duration(qualtrics_text, min_duration = 100)

# Remove preview data first
df <- qualtrics_text \%>\%
  exclude_preview() \%>\%
  mark_duration()

# Mark only for durations slower than 800 seconds
df <- qualtrics_text \%>\%
  exclude_preview() \%>\%
  mark_duration(max_duration = 800)
}
\seealso{
Other duration functions: 
\code{\link{check_duration}()},
\code{\link{exclude_duration}()}

Other mark functions: 
\code{\link{mark_duplicates}()},
\code{\link{mark_ip}()},
\code{\link{mark_location}()},
\code{\link{mark_preview}()},
\code{\link{mark_progress}()},
\code{\link{mark_resolution}()}
}
\concept{duration functions}
\concept{mark functions}
