% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_progress.R
\name{mark_progress}
\alias{mark_progress}
\title{Mark survey progress}
\usage{
mark_progress(.data, id_col = "ResponseId", ...)
}
\arguments{
\item{.data}{Data frame or tibble (preferably exported from Qualtrics).}

\item{id_col}{Column name for unique row ID (e.g., participant).}

\item{...}{Inherit parameters from check function.}
}
\value{
An object of the same type as \code{.data} that includes a column marking rows
that have incomplete progress.
For a function that excludes these rows, use \code{\link[=exclude_progress]{exclude_progress()}}.
For a function that marks these rows, use \code{\link[=mark_progress]{mark_progress()}}.
}
\description{
The \code{mark_progress()} function creates a column labeling
rows that have incomplete progress.
The function is written to work with data from
\href{https://qualtrics.com}{Qualtrics} surveys.
}
\details{
Default column names are set based on output from the
\code{\link[qualtRics:fetch_survey]{qualtRics::fetch_survey()}}.
The default requires 100\% completion, but lower levels of completion
maybe acceptable and can be allowed by specifying the \code{min_progress} argument.
The finished column in Qualtrics can be a numeric or character vector
depending on whether it is exported as choice text or numeric values.
This function works for both.

The function outputs to console a message about the number of rows
that have incomplete progress.
}
\examples{
# Mark rows with incomplete progress
data(qualtrics_text)
df <- mark_progress(qualtrics_text)

# Remove preview data first
df <- qualtrics_text \%>\%
  exclude_preview() \%>\%
  mark_progress()

# Include a lower acceptable completion percentage
df <- qualtrics_numeric \%>\%
  exclude_preview() \%>\%
  mark_progress(min_progress = 98)

# Do not print rows to console
df <- qualtrics_text \%>\%
  exclude_preview() \%>\%
  mark_progress(print_tibble = FALSE)

# Do not print message to console
df <- qualtrics_text \%>\%
  exclude_preview() \%>\%
  mark_progress(quiet = TRUE)

}
\seealso{
Other progress functions: 
\code{\link{check_progress}()}

Other mark functions: 
\code{\link{mark_duplicates}()},
\code{\link{mark_duration}()},
\code{\link{mark_ip}()},
\code{\link{mark_location}()},
\code{\link{mark_preview}()},
\code{\link{mark_resolution}()}
}
\concept{mark functions}
\concept{progress functions}
