% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mark_rows}
\alias{mark_rows}
\title{Return marked rows}
\usage{
mark_rows(x, filtered_data, id_col, exclusion_type)
}
\arguments{
\item{x}{Original data.}

\item{filtered_data}{Data to be excluded.}

\item{id_col}{Column name for unique row ID (e.g., participant).}

\item{exclusion_type}{Column name for exclusion column.}
}
\description{
Create new column marking rows that meet exclusion criteria.
\emph{This function is not exported.}
}
\keyword{internal}
