% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_data}
\alias{print_data}
\title{Print data to console}
\usage{
print_data(x, print)
}
\arguments{
\item{x}{Data set to print or not}

\item{print}{Logical indicating whether to print returned tibble to
console.}
}
\description{
Prints the data to the console. \emph{This function is not exported.}
}
