% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_exclusion}
\alias{print_exclusion}
\title{Print number of excluded rows}
\usage{
print_exclusion(remaining_data, x, msg)
}
\arguments{
\item{remaining_data}{Data after removing exclusions.}

\item{x}{Original data before removing exclusions.}

\item{msg}{Text to describe what types of rows were excluded.}
}
\description{
Prints a message to the console with the number of excluded rows.  \emph{This function is not exported.}
}
\keyword{internal}
