% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_data.R
\docType{data}
\name{qualtrics_anonymous}
\alias{qualtrics_anonymous}
\title{Example numeric metadata imported with \code{qualtRics::fetch_survey()} from
simulated Qualtrics study with anonymized data}
\format{
A data frame with 100 rows and 10 variables:
\describe{
\item{StartDate}{date and time data collection started, in ISO 8601 format}
\item{EndDate}{date and time data collection ended, in ISO 8601 format}
\item{Status}{numeric flag for preview (1) vs. implemented survey (0)
entries}
\item{Progress}{percentage of survey completed}
\item{Duration (in seconds)}{duration of time required to complete survey,
in seconds}
\item{Finished}{numeric flag for whether survey was completed (1) or
progress was < 100 (0)}
\item{RecordedDate}{date and time survey was recorded, in ISO 8601 format}
\item{ResponseId}{random ID for participants}
\item{UserLanguage}{language set in Qualtrics}
\item{Q2}{response to question about whether the user liked the survey
(1 = Yes, 0 = No)}
}
}
\usage{
qualtrics_anonymous
}
\description{
A dataset containing the metadata from a standard Qualtrics survey with
anonymized data collection and exported with "Use numeric values". This data
do not include IP addresses, location, or computer information. The data
were imported using
\href{https://docs.ropensci.org/qualtRics/reference/fetch_survey.html}{\code{qualtRics::fetch_survey()}}.
These data were randomly generated using \href{https://hrbrmstr.github.io/iptools/reference/ip_random.html}{iptools::ip_random()} and
\href{https://cran.r-project.org/package=rgeolocate}{rgeolocate::ip2location()} functions.
}
\seealso{
Other data: 
\code{\link{qualtrics_fetch}},
\code{\link{qualtrics_fetch2}},
\code{\link{qualtrics_numeric}},
\code{\link{qualtrics_raw}},
\code{\link{qualtrics_text}}
}
\concept{data}
\keyword{datasets}
