% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_text.R
\docType{data}
\name{qualtrics_text}
\alias{qualtrics_text}
\title{Example text-based metadata from simulated Qualtrics study}
\format{
A data frame with 100 rows and 16 variables:
\describe{
\item{StartDate}{date and time data collection started, in ISO 8601 format}
\item{EndDate}{date and time data collection ended, in ISO 8601 format}
\item{Status}{flag for preview (Survey Preview) vs. implemented survey (IP Address) entries}
\item{IPAddress}{participant IP address (truncated for anonymity)}
\item{Progress}{percentage of survey completed}
\item{Duration (in seconds)}{duration of time required to complete survey, in seconds}
\item{Finished}{logical for whether survey was completed (TRUE) or progress was < 100 (FALSE)}
\item{RecordedDate}{date and time survey was recorded, in ISO 8601 format}
\item{ResponseId}{random ID for participants}
\item{LocationLatitude}{latitude geolocated from IP address}
\item{LocationLongitude}{longitude geolocated from IP address}
\item{UserLanguage}{language set in Qualtrics}
\item{Browser}{user web browser type}
\item{Version}{user web browser version}
\item{Operation System}{user operating system}
\item{Resolution}{user screen resolution}
}
}
\usage{
qualtrics_text
}
\description{
A dataset containing the metadata from a standard Qualtrics survey with
browser metadata collected and exported with "Use choice text".
These data were randomly generated using #' \link[iptools:ip_random]{iptools::ip_random} and
\link[rgeolocate:ip2location]{rgeolocate::ip2location} functions.
}
\keyword{datasets}
