% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_columns}
\alias{validate_columns}
\title{Check number, names, and type of columns}
\usage{
validate_columns(x, column)
}
\arguments{
\item{x}{Data set.}

\item{column}{Name of column argument to check.}

\item{col_num}{Expected number of columns.}
}
\description{
Determines whether the correct number and names of columns were specified
as arguments to the functions. \emph{This function is not exported.}
}
\keyword{internal}
