% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicates.R
\name{check_duplicates}
\alias{check_duplicates}
\title{Title}
\usage{
check_duplicates(
  .data,
  ip_col = "IPAddress",
  location_col = c("LocationLatitude", "LocationLongitude"),
  check_ip = TRUE,
  check_location = TRUE,
  include_na = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
