% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duration.R
\name{check_duration}
\alias{check_duration}
\title{Title}
\usage{
check_duration(
  .data,
  min_duration = NULL,
  max_duration = NULL,
  duration_col = "Duration (in seconds)",
  print_tibble = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
