% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_exclusions.R
\name{collapse_exclusions}
\alias{collapse_exclusions}
\title{Unite multiple exclusion columns into single column}
\usage{
collapse_exclusions(
  x,
  exclusion_types = c("duplicates", "duration", "ip", "location", "preview",
    "progress", "resolution"),
  separator = ",",
  remove = TRUE
)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{collapse_exclusions()} was renamed to \code{\link[=unite_exclusions]{unite_exclusions()}} to create a more
consistent API with tidyverse's \code{unite()} function.
}
\keyword{internal}
