% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_duration.R
\name{exclude_duration}
\alias{exclude_duration}
\title{Title}
\usage{
exclude_duration(
  .data,
  duration_col = "Duration (in seconds)",
  min = NULL,
  max = NULL,
  id_col = "ResponseId",
  quiet = FALSE
)
}
\arguments{
\item{quiet}{}
}
\value{

}
\description{
Title
}
