% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excluder-package.R
\docType{package}
\name{excluder-package}
\alias{excluder}
\alias{excluder-package}
\title{excluder: Checks for Exclusion Criteria in Online Data}
\description{
Data that are collected through online sources such as Mechanical Turk may require excluding data because of IP address duplication, geolocation, or completion duration. This package facilitates exclusion of these data for Qualtrics datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jeffreyrstevens.github.io/excluder/}
  \item \url{https://github.com/jeffreyrstevens/excluder/}
  \item Report bugs at \url{https://github.com/jeffreyrstevens/excluder/issues/}
}

}
\author{
\strong{Maintainer}: Jeffrey R. Stevens \email{jeffrey.r.stevens@gmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID})

Other contributors:
\itemize{
  \item Joseph O'Brien (\href{https://orcid.org/0000-0001-9851-5077}{ORCID}) [reviewer]
  \item Julia Silge \email{julia.silge@gmail.com} (\href{https://orcid.org/0000-0002-3671-836X}{ORCID}) [reviewer]
}

}
\keyword{internal}
