% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_exclusions.R
\name{mark_exclusions}
\alias{mark_exclusions}
\title{Title}
\usage{
mark_exclusions(
  .data,
  exclusion_types = c("duplicates", "duration", "ip", "language", "location",
    "preview", "progress", "resolution"),
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
