% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtrics_data.R
\docType{data}
\name{qualtrics_fetch}
\alias{qualtrics_fetch}
\title{Example numeric metadata imported with qualtRics::fetch_survey() from
simulated Qualtrics study}
\format{
A data frame with 100 rows and 17 variables:
\describe{
\item{StartDate}{date and time data collection started, in ISO 8601 format}
\item{EndDate}{date and time data collection ended, in ISO 8601 format}
\item{Status}{numeric flag for preview (1) vs. implemented survey (0)
entries}
\item{IPAddress}{participant IP address (truncated for anonymity)}
\item{Progress}{percentage of survey completed}
\item{Duration (in seconds)}{duration of time required to complete survey,
in seconds}
\item{Finished}{numeric flag for whether survey was completed (1) or
progress was < 100 (0)}
\item{RecordedDate}{date and time survey was recorded, in ISO 8601 format}
\item{ResponseId}{random ID for participants}
\item{LocationLatitude}{latitude geolocated from IP address}
\item{LocationLongitude}{longitude geolocated from IP address}
\item{UserLanguage}{language set in Qualtrics}
\item{Q1_Browser}{user web browser type}
\item{Q1_Version}{user web browser version}
\item{Q1_Operating System}{user operating system}
\item{Q1_Resolution}{user screen resolution}
\item{Q2}{response to question about whether the user liked the survey
(1 = Yes, 0 = No)}
}
}
\usage{
qualtrics_fetch
}
\description{
A dataset containing the metadata from a standard Qualtrics survey with
browser metadata collected and exported with "Use numeric values". The data
were imported using \code{\link[qualtRics:fetch_survey]{qualtRics::fetch_survey()}} and then the secondary labels
were assigned as column names with \code{\link[sjlabelled:get_label]{sjlabelled::get_label()}}.
These data were randomly generated using \code{\link[iptools:ip_random]{iptools::ip_random()}} and
\href{https://cran.r-project.org/package=rgeolocate}{rgeolocate::ip2location()} functions.
}
\seealso{
Other data: 
\code{\link{qualtrics_numeric}},
\code{\link{qualtrics_raw}},
\code{\link{qualtrics_text}}
}
\concept{data}
\keyword{datasets}
